+++ packages/attributes/src/attributes.ts
@@ -1,10 +1,13 @@
 import {
+  attachExternalStylesheets,
   ATTRIBUTES,
+  fetchPage,
   type FinsweetAttributeKey,
   type FinsweetAttributesCallback,
   waitDOMReady,
 } from '@finsweet/attributes-utils';
 
+import { version } from '../package.json';
 import { loadAttribute } from './load';
 
 const ATTRIBUTES_ATTRIBUTE_PREFIX = 'fs-attributes';
@@ -31,9 +34,14 @@ const init = () => {
 
   // Init Attributes object
   window.FinsweetAttributes = {
+    version,
     scripts,
     modules: {},
     process: new Set<FinsweetAttributeKey>(),
+    utils: {
+      fetchPage,
+      attachExternalStylesheets,
+    },
 
     load: initAttribute,