+++ packages/list/src/filter/standard/conditions.ts
@@ -1,14 +1,24 @@
 import {
   type FormField,
   type FormFieldType,
+  getCheckboxGroup,
   getFormFieldValue,
+  getRadioGroupInputs,
   isFormField,
+  isString,
   setFormFieldValue,
 } from '@finsweet/attributes-utils';
 
 import type { List } from '../../components/List';
-import { CUSTOM_VALUE_ATTRIBUTE, getAttribute, getSettingSelector, hasAttributeValue } from '../../utils/selectors';
+import {
+  CUSTOM_VALUE_ATTRIBUTE,
+  getAttribute,
+  getSettingSelector,
+  getSplitSeparator,
+  hasAttributeValue,
+} from '../../utils/selectors';
 import type { FiltersCondition, FiltersGroup } from '../types';
+import { splitValue } from '../utils';
 
 /**
  * @returns The value of a given form field.
@@ -22,13 +32,20 @@ export const getConditionData = (formField: FormField, fieldKey: string, interac
   const op = getConditionOperator(formField);
   const id = `${fieldKey}_${op}`;
 
-  const customTagField = getAttribute(formField, 'tagfield');
+  const tagCustomField = getAttribute(formField, 'tagfield');
+  const tagCustomValues = getTagCustomValues(formField);
+  const tagValuesDisplay = getAttribute(formField, 'tagvalues', { filterInvalid: true });
   const filterMatch = getAttribute(formField, 'filtermatch', { filterInvalid: true });
   const fieldMatch = getAttribute(formField, 'fieldmatch', { filterInvalid: true });
   const fuzzyThreshold = getAttribute(formField, 'fuzzy');
   const showTag = !hasAttributeValue(formField, 'showtag', 'false');
+  const splitSeparator = getSplitSeparator(formField);
 
-  const value = getFormFieldValue(formField, CUSTOM_VALUE_ATTRIBUTE);
+  let value = getFormFieldValue(formField, CUSTOM_VALUE_ATTRIBUTE);
+
+  if (isString(value) && splitSeparator) {
+    value = splitValue(value, splitSeparator);
+  }
 
   return {
     id,
@@ -40,7 +57,9 @@ export const getConditionData = (formField: FormField, fieldKey: string, interac
     fieldMatch,
     fuzzyThreshold,
     interacted,
-    customTagField,
+    tagCustomField,
+    tagCustomValues,
+    tagValuesDisplay,
     showTag,
   };
 };
@@ -120,4 +139,62 @@ export const getStandardFiltersGroup = (list: List, form: HTMLFormElement, group
   list.readingFilters = false;
 
   return group;
+};
+
+/**
+ * @returns A map of tag values for a given form field.
+ * @param formField
+ */
+const getTagCustomValues = (formField: FormField): Map<string, string> | undefined => {
+  let tagCustomValues: Map<string, string> | undefined;
+
+  const type = formField.type as FormFieldType;
+
+  switch (type) {
+    case 'checkbox': {
+      // Group
+      const groupCheckboxes = getCheckboxGroup(formField.name, formField.form, CUSTOM_VALUE_ATTRIBUTE);
+      if (groupCheckboxes?.length) {
+        for (const checkbox of groupCheckboxes) {
+          const checkboxValue = checkbox.getAttribute(CUSTOM_VALUE_ATTRIBUTE) ?? checkbox.value;
+          if (!checkboxValue) continue;
+
+          const tagValue = getAttribute(checkbox, 'tagvalue');
+          if (!tagValue) continue;
+
+          tagCustomValues ||= new Map<string, string>();
+          tagCustomValues.set(checkboxValue, tagValue);
+        }
+
+        break;
+      }
+
+      // Single
+      const tagValue = getAttribute(formField, 'tagvalue');
+      if (!tagValue) break;
+
+      tagCustomValues = new Map<string, string>([['true', tagValue]]);
+      break;
+    }
+
+    case 'radio': {
+      const groupRadios = getRadioGroupInputs(formField);
+      if (!groupRadios?.length) break;
+
+      for (const radio of groupRadios) {
+        const radioValue = radio.getAttribute(CUSTOM_VALUE_ATTRIBUTE) ?? radio.value;
+        if (!radioValue) continue;
+
+        const tagValue = getAttribute(radio, 'tagvalue');
+        if (!tagValue) continue;
+
+        tagCustomValues ||= new Map<string, string>();
+        tagCustomValues.set(radioValue, tagValue);
+      }
+
+      break;
+    }
+  }
+
+  return tagCustomValues;
 };