+++ packages/list/src/init.ts
@@ -2,7 +2,7 @@ import { type FinsweetAttributeInit, isNotEmpty, waitWebflowReady } from '@finsw
 
 import { createListInstance, initList } from './factory';
 import { getCMSElementSelector } from './utils/dom';
-import { queryAllElements } from './utils/selectors';
+import { getInstance, queryAllElements } from './utils/selectors';
 
 /**
  * Inits the attribute.
@@ -14,12 +14,29 @@ export const init: FinsweetAttributeInit = async () => {
   const lists = listElements
     .map((listElement) => {
       const listSelector = getCMSElementSelector('list');
-      const parentList = listElement.parentElement?.closest(listSelector);
 
-      // TODO: We don't support nested CMS lists for now,
-      // but we may want to revisit this in the future
-      const isNestedList = !!parentList && listElement !== parentList;
-      if (isNestedList) return;
+      const parentLists: Element[] = [];
+
+      let parentList = listElement.parentElement?.closest(listSelector);
+      while (parentList) {
+        parentLists.push(parentList);
+        parentList = parentList.parentElement?.closest(listSelector);
+      }
+
+      const isNestedList = parentLists.length > 0;
+      if (isNestedList) {
+        const listInstance = getInstance(listElement);
+
+        const parentListHasInstance = parentLists.some((parentList) => {
+          const parentListInstance = getInstance(parentList);
+          return listInstance === parentListInstance;
+        });
+
+        // Nested lists can only be initialized if
+        // they have a different instance than their parent lists
+        // to avoid unexpected behaviors
+        if (parentListHasInstance) return;
+      }
 
       return createListInstance(listElement);
     })