+++ packages/list/src/load/index.ts
@@ -42,7 +42,11 @@ export const initListLoading = (list: List, mode: LoadMode) => {
 
       const animations = list.wrapperElement.getAnimations({ subtree: true });
 
-      await Promise.all(animations.map((a) => a.finished));
+      try {
+        await Promise.all(animations.map((a) => a.finished));
+      } catch {
+        //
+      }
     }
 
     return items;
@@ -54,8 +58,8 @@ export const initListLoading = (list: List, mode: LoadMode) => {
     return items;
   });
 
-  return () => {
-    loadModeCleanup?.();
+  return async () => {
+    (await loadModeCleanup)?.();
     beforeRenderHookCleanup();
     afterRenderHookCleanup();
     elementsCleanup();