+++ packages/list/src/load/pagination.ts
@@ -24,7 +24,7 @@ import { loadPaginatedCMSItems } from './load';
  *
  * @returns A callback to destroy the event listeners.
  */
-export const initPaginationMode = (list: List) => {
+export const initPaginationMode = async (list: List) => {
   const { currentPage, itemsPerPage, paginationWrapperElement } = list;
   if (!paginationWrapperElement) return;
 
@@ -36,6 +36,7 @@ export const initPaginationMode = (list: List) => {
     const end = start + $itemsPerPage;
 
     const paginatedItems = items.slice(start, end);
+
     return paginatedItems;
   });
 
@@ -44,12 +45,12 @@ export const initPaginationMode = (list: List) => {
     () => {
       // We reset currentPage to 1 when filtering and sorting,
       // we don't want to trigger pagination again when that happens
-      if (list.currentHook) return;
-
-      list.triggerHook('pagination', { scrollToAnchor: true });
+      if (!list.triggeredHook) {
+        list.triggerHook('pagination', { scrollToAnchor: true });
+      }
 
       if (list.showQuery) {
-        list.setSearchParam('page', currentPage.value.toString());
+        list.setSearchParam('page', currentPage.value.toString(), { useSearchParamsPrefix: true });
       }
     },
     {}
@@ -58,6 +59,11 @@ export const initPaginationMode = (list: List) => {
   const paginateCleanup = watch(list.itemsPerPage, () => {
     list.triggerHook('pagination', { scrollToAnchor: true });
   });
+
+  // Init items load
+  loadPaginatedCMSItems(list);
+
+  await list.loadingSearchParamsData;
 
   // Get settings
   const [pageBoundary, pageBoundaryCleanup] = getBreakpointSetting(list, 'pageboundary');
@@ -95,9 +101,6 @@ export const initPaginationMode = (list: List) => {
       );
     }
   }
-
-  // Init items load
-  loadPaginatedCMSItems(list);
 
   // Handle pagination elements
   const paginationWrapperCleanup = handlePaginationWrapper(list);