+++ packages/list/src/tabs/index.ts
@@ -1,6 +1,6 @@
-import { ARIA_ROLE_KEY, TABS_CSS_CLASSES } from '@finsweet/attributes-utils';
+import { ARIA_ROLE_KEY, cloneNode, TABS_CSS_CLASSES } from '@finsweet/attributes-utils';
 
-import type { List, ListItem } from '../components';
+import type { List } from '../components';
 import { queryElement } from '../utils/selectors';
 
 /**
@@ -46,17 +46,18 @@ const initListTab = (list: List, tabsReference: HTMLElement) => {
 
   // Store rendered items
   const renderedItems = new Map<
-    ListItem,
+    string,
     {
       tabLink: HTMLElement;
       tabPane: HTMLElement;
     }
   >();
 
-  list.addHook('render', (items = []) => {
+  list.addHook('beforeRender', (items = []) => {
     for (const item of items) {
-      if (renderedItems.has(item)) continue;
+      if (renderedItems.has(item.id)) continue;
 
+      item.currentIndex = undefined;
       item.element.removeAttribute(ARIA_ROLE_KEY);
 
       const tabLink = document.createElement('div');
@@ -66,7 +67,10 @@ const initListTab = (list: List, tabsReference: HTMLElement) => {
       tabPane.setAttribute('class', tabPaneCSS);
 
       let tabLinkContent = queryElement('tab-link', { scope: item.element });
-      if (!tabLinkContent) {
+
+      if (tabLinkContent) {
+        tabLinkContent = cloneNode(tabLinkContent);
+      } else {
         tabLinkContent = document.createElement('div');
         tabLinkContent.innerHTML = /* html */ `Missing <strong>fs-list-element="tab-link"</strong>`;
       }
@@ -76,15 +80,15 @@ const initListTab = (list: List, tabsReference: HTMLElement) => {
       tabPane.appendChild(item.element);
       tabsContent.appendChild(tabPane);
 
-      renderedItems.set(item, { tabLink, tabPane });
+      renderedItems.set(item.id, { tabLink, tabPane });
     }
 
-    for (const [item, { tabLink, tabPane }] of renderedItems) {
-      if (items.includes(item)) continue;
+    for (const [itemId, { tabLink, tabPane }] of renderedItems) {
+      if (items.some((item) => item.id === itemId)) continue;
 
       tabLink?.remove();
       tabPane?.remove();
-      renderedItems.delete(item);
+      renderedItems.delete(itemId);
     }
 
     [...renderedItems.values()].forEach(({ tabLink, tabPane }, index) => {
@@ -93,5 +97,11 @@ const initListTab = (list: List, tabsReference: HTMLElement) => {
       tabLink.dataset.wTab = dataset;
       tabPane.dataset.wTab = dataset;
     });
+
+    list.renderedItems.clear();
+
+    return [];
   });
+
+  list.triggerHook('beforeRender');
 };