+++ packages/list/src/utils/constants.ts
@@ -551,6 +551,12 @@ export const SETTINGS = {
   value: { key: 'value' },
 
   /**
+   * Defines a split separator for a filter value.
+   * If set to `true`, it will use a space as a separator.
+   */
+  split: { key: 'split' },
+
+  /**
    * Defines a list instance where the list should be combined with.
    */
   combine: { key: 'combine' },
@@ -569,6 +575,16 @@ export const SETTINGS = {
    * Defines a custom display value for a field in a tag.
    */
   tagfield: { key: 'tagfield' },
+
+  /**
+   * Defines a custom display value for a value in a tag.
+   */
+  tagvalue: { key: 'tagvalue' },
+
+  /**
+   * Defines how to render tag values.
+   */
+  tagvalues: { key: 'tagvalues', values: ['combined', 'separate'], defaultValue: 'combined' },
 
   /**
    * Defines if the tag for a specific filter should be displayed.
@@ -633,6 +649,12 @@ export const SETTINGS = {
    * Defines if Webflow's ix2 should be reset when rendering the list items.
    */
   resetix: { key: 'resetix', values: ['true'] },
+
+  /**
+   * Alternative to `fs-list-element="facet-count"`
+   * for those elements that already have an `fs-list-element` attribute.
+   */
+  facetcount: { key: 'facetcount' },
 } as const satisfies AttributeSettings;
 
 export const BREAKPOINTS_INDEX: { [key in WebflowBreakpoint]: number } = {
@@ -720,7 +742,7 @@ export const ALLOWED_DYNAMIC_FIELD_TYPES: Record<
 
 export const RENDER_INDEX_CSS_VARIABLE = `--fs-${LIST_ATTRIBUTE}-renderindex`;
 
-window.CSS.registerProperty({
+window.CSS?.registerProperty?.({
   name: RENDER_INDEX_CSS_VARIABLE,
   syntax: '<number>',
   inherits: false,