+++ packages/list/src/utils/selectors.ts
@@ -16,3 +16,14 @@ export const {
 } = generateSelectors(LIST_ATTRIBUTE, ELEMENTS, SETTINGS);
 
 export const CUSTOM_VALUE_ATTRIBUTE = getSettingAttributeName('value');
+
+/**
+ * @returns The split separator for the element.
+ * @param element
+ */
+export const getSplitSeparator = (element: Element) => {
+  const rawSplitSeparator = getAttribute(element, 'split');
+
+  const splitSeparator = rawSplitSeparator === 'true' ? ' ' : rawSplitSeparator;
+  return splitSeparator;
+};