+++ packages/utils/src/types/core.ts
@@ -1,5 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import type * as ATTRIBUTES from '../constants/attributes';
+import { attachExternalStylesheets, fetchPage } from '../helpers';
 
 export type FinsweetAttributeKey = (typeof ATTRIBUTES)[keyof typeof ATTRIBUTES];
 
@@ -34,6 +35,11 @@ export type FinsweetAttributes = {
   destroy?: () => void;
 
   /**
+   * Defines the Attributes library version.
+   */
+  version: string;
+
+  /**
    * Contains access to each Attribute solution.
    */
   modules: {
@@ -50,6 +56,11 @@ export type FinsweetAttributes = {
    * Can be multiple as the user might import the library multiple times.
    */
   scripts: HTMLScriptElement[];
+
+  utils: {
+    fetchPage: typeof fetchPage;
+    attachExternalStylesheets: typeof attachExternalStylesheets;
+  };
 };
 
 export type FinsweetAttributesCallback = [FinsweetAttributeKey, (value: any) => void];