+++ packages/docs/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @finsweet/attributes-docs
 
+## 2.0.16
+
+### Patch Changes
+
+- Updated dependencies [894db29]
+  - @finsweet/attributes-utils@0.1.4
+
+## 2.0.15
+
+### Patch Changes
+
+- Updated dependencies [ca570e1]
+  - @finsweet/attributes-utils@0.1.3
+
+## 2.0.14
+
+### Patch Changes
+
+- Updated dependencies [4e55c67]
+  - @finsweet/attributes-utils@0.1.2
+
 ## 2.0.13
 
 ### Patch Changes