+++ packages/sliderdots/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @finsweet/attributes-sliderdots
 
+## 1.8.11
+
+### Patch Changes
+
+- Updated dependencies [894db29]
+  - @finsweet/attributes-utils@0.1.4
+
+## 1.8.10
+
+### Patch Changes
+
+- Updated dependencies [ca570e1]
+  - @finsweet/attributes-utils@0.1.3
+
+## 1.8.9
+
+### Patch Changes
+
+- Updated dependencies [4e55c67]
+  - @finsweet/attributes-utils@0.1.2
+
 ## 1.8.8
 
 ### Patch Changes