+++ packages/smartlightbox/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @finsweet/attributes-smartlightbox
 
+## 1.7.11
+
+### Patch Changes
+
+- Updated dependencies [894db29]
+  - @finsweet/attributes-utils@0.1.4
+
+## 1.7.10
+
+### Patch Changes
+
+- Updated dependencies [ca570e1]
+  - @finsweet/attributes-utils@0.1.3
+
+## 1.7.9
+
+### Patch Changes
+
+- Updated dependencies [4e55c67]
+  - @finsweet/attributes-utils@0.1.2
+
 ## 1.7.8
 
 ### Patch Changes