+++ packages/toc/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @finsweet/attributes-toc
 
+## 1.2.16
+
+### Patch Changes
+
+- Updated dependencies [894db29]
+  - @finsweet/attributes-utils@0.1.4
+
+## 1.2.15
+
+### Patch Changes
+
+- 67c32c7: fix: add a fallback for TOC item id's when the language cannot be slugified
+
+## 1.2.14
+
+### Patch Changes
+
+- Updated dependencies [ca570e1]
+  - @finsweet/attributes-utils@0.1.3
+
+## 1.2.13
+
+### Patch Changes
+
+- Updated dependencies [4e55c67]
+  - @finsweet/attributes-utils@0.1.2
+
 ## 1.2.12
 
 ### Patch Changes